fixed a type storage issue that was discarding flags greater than 16 bits

+++ vterm.h
@@ -74,34 +74,35 @@ typedef chtype          attr_t;
 
 #define LIBVTERM_VERSION        "5.14"
 
-#define VTERM_FLAG_RXVT         (1 << 0)    //  masquerade as rxvt (default)
-#define VTERM_FLAG_VT100        (1 << 1)    //  masquerade as vt100
-#define VTERM_FLAG_XTERM        (1 << 2)    //  masquerade as xterm
-#define VTERM_FLAG_XTERM_256    (1 << 3)    //  masquerade as xterm-256
-#define VTERM_FLAG_LINUX        (1 << 4)    //  masquerade as linux
+#define VTERM_FLAG_RXVT         (1UL << 0)      //  emulate rxvt
+#define VTERM_FLAG_VT100        (1UL << 1)      //  emulate vt100
+#define VTERM_FLAG_XTERM        (1UL << 2)      //  emulate xterm (default)
+#define VTERM_FLAG_XTERM_256    (1UL << 3)      //  emulate xterm-256
+#define VTERM_FLAG_LINUX        (1UL << 4)      //  emulate linux
 
-#define VTERM_FLAG_NOPTY        (1 << 8)    /*
-                                                skip all the fd and pty stuff.
-                                                just render input args byte
-                                                stream to a buffer.
-                                            */
+#define VTERM_FLAG_NOPTY        (1UL << 8)      /*
+                                                    skip all the fd and pty
+                                                    stuff.  just render input
+                                                    args byte stream to a
+                                                    buffer.
+                                                */
 
-#define VTERM_FLAG_NOCURSES     (1 << 9)    /*
-                                                skip the curses WINDOW stuff.
-                                                return the char cell array if
-                                                required.
-                                            */
+#define VTERM_FLAG_NOCURSES     (1UL << 9)      /*
+                                                    skip the curses WINDOW
+                                                    stuff.  return the char
+                                                    cell array if required.
+                                                */
 
-#define VTERM_FLAG_C16          (1 << 15)   /*
-                                                emulate aixterm 16 colors
-                                                with mapped colors.
-                                            */
+#define VTERM_FLAG_C16          (1UL << 15)     /*
+                                                    emulate aixterm 16 colors
+                                                    with mapped colors.
+                                                */
 
-#define VTERM_FLAG_DUMP         (1 << 31)   /*
-                                                tell libvterm to write
-                                                stream data to a dump file
-                                                for debugging.
-                                            */
+#define VTERM_FLAG_DUMP         (1UL << 31)     /*
+                                                    tell libvterm to write
+                                                    stream data to a dump file
+                                                    for debugging.
+                                                */
 
 
 // #define VCELL_TYPE_SIMPLE       0
@@ -204,7 +205,7 @@ vterm_t*        vterm_alloc(void);
     @return:        a handle to a newly alloc'd vterm object if not provided.
 */
 vterm_t*        vterm_init(vterm_t *vterm, uint16_t width, uint16_t height,
-                    uint16_t flags);
+                    uint32_t flags);
 
 /*
     convenience macro for alloc-ing a ready to use terminal object.
